Create Project1
Read lines from a file that is supposed to contain words. If the word read is valid (has only letters) put the word into an array (otherwise print it to the console). Once the file is completely read, put the contents of the array into a TextArea and display the TextArea in the left half of a GUI (JFrame). Then sort the array using Selection Sort, and put the resulting array into another TextArea which should then be displayed in the right half of the GUI.